+++ src/main/java/org/chemtrovina/iom_systemscan/service/InvoiceService.java
@@ -5,9 +5,9 @@ import org.chemtrovina.iom_systemscan.model.Invoice;
 import java.util.List;
 
 public interface InvoiceService {
-    Invoice getInvoiceById(Long id);
+    Invoice getInvoiceById(int id);
     List<Invoice> getAllInvoices();
     void createInvoice(Invoice invoice);
     void updateInvoice(Invoice invoice);
-    void deleteInvoice(Long id);
+    void deleteInvoice(int id);
 }
